refactor: extract lakehouse code into dedicated files
New files:
- activities/metadata_extraction/lakehouse.py
  All lakehouse implementation: submit_and_poll_mdlh_load,
  convert_raw_parquet_to_jsonl

- workflows/metadata_extraction/lakehouse.py
  LakehouseLoadMixin with load_raw_to_lakehouse,
  load_transformed_to_lakehouse, _submit_lakehouse_load,
  resolve_iceberg_table

Existing files now only contain thin delegation:
- base.py: activity methods delegate to lakehouse.py functions
- sql.py: same delegation for SQL activity class
- __init__.py: MetadataExtractionWorkflow inherits LakehouseLoadMixin
